feat: add config options to assign cost to revealing soldier stats
Adds config options that allow users to assign a cost to revealing Black
Market soldier stats. Both the resource type (Intel or Supplies) as well
as the amount to pay can be configured via the mod settings. It's also
possible to always show the soldier stats without having to pay a price,
which is also the default behaviour.